requirements.txt
Django==5.0.4
django-elasticsearch-dsl==7.4
django-elasticsearch-dsl-drf==0.22.5
djangorestframework==3.15.1
elasticsearch==7.17.2
elasticsearch-dsl==7.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0

